fix(protocol): set int as return value of evaluated decision version getter
The decision version getter in the EvaluatedDecisionValue interface returns a Long. This is inconsistent since the version property is of int type in the EvaluatedDecisionRecord and the DecisionEvaluationRecord. Furthermore, the same decision version getter for the DecisionEvaluatonRecord returns an int type. This commit makes all of the decision version getters consistent.

Due to the update, the generated ImmutableEvaluatedDecisionValue is updated, and requires an update of the revapi.json. The breaking change is acceptable since the decision version property is unlikely to surpass MAX_INT. Furthermore, in many cases the JVM automatically converts Long to Int type.

+++ protocol/src/main/java/io/camunda/zeebe/protocol/record/value/EvaluatedDecisionValue.java
@@ -45,7 +45,7 @@ public interface EvaluatedDecisionValue extends RecordValue {
   /**
    * @return the version of the evaluated decision
    */
-  long getDecisionVersion();
+  int getDecisionVersion();
 
   /**
    * @return the type of the evaluated decision
